+++ src/test-run/commands/service.js
@@ -1,0 +1,14 @@
+import TYPE from './type';
+
+// Commands
+export class PrepareBrowserManipulationCommand {
+    constructor () {
+        this.type = TYPE.prepareBrowserManipulation;
+    }
+}
+
+export class TestDoneCommand {
+    constructor () {
+        this.type = TYPE.testDone;
+    }
+}
